Error fixtures and code formatting #40

+++ data-access/src/main/java/com/healthplus/dataaccess/repo/PatientRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.healthplus.dataaccess.domain.Patient;
 
-public interface PatientRepository extends JpaRepository<Patient, Integer> {
+public interface PatientRepository extends JpaRepository<Patient, Long> {
 
-	@Query(value="SELECT * FROM patient WHERE email = ?1", nativeQuery=true)
+	@Query(value = "SELECT * FROM patient WHERE email = ?1", nativeQuery = true)
 	public Optional<Patient> getPatientByEmail(String email);
 
-	@Query(value="SELECT * FROM patient WHERE contact = ?1", nativeQuery=true)
+	@Query(value = "SELECT * FROM patient WHERE contact = ?1", nativeQuery = true)
 	public Optional<Patient> getPatientByContact(Long contact);
 }
